Fixed unit tests and a couple of fixes more

+++ etl/aro/create_aro_businesses.sql
@@ -4,7 +4,7 @@ DROP TABLE IF EXISTS aro.businesses;
 
 CREATE TABLE aro.businesses
 (
-	id bigint,
+	id serial,
 	location_id bigint,
 	industry_id int,
 	name varchar,
